Update directories and practice calling API
