+++ edmtrain-reskin/package.json
@@ -12,10 +12,6 @@
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
-
-  "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^1.0.10"
-  },
   
   "scripts": {
     "start": "react-scripts start",
